feat: implement remove product service

+++ src/category/category.controller.ts
@@ -3,11 +3,15 @@ import { CategoryService } from './category.service';
 import { CreateCategoryDto } from './dto/create-category.dto';
 import { UpdateCategoryDto } from './dto/update-category.dto';
 import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
+import { ProductService } from 'src/product/product.service';
 
 @ApiTags('Category')
 @Controller('category')
 export class CategoryController {
-  constructor(private readonly categoryService: CategoryService) {}
+  constructor(
+    private readonly categoryService: CategoryService,
+    private readonly productService: ProductService
+  ) {}
 
   @Post()
   @ApiOperation({ summary: 'Create a new category' })
@@ -40,7 +44,11 @@ export class CategoryController {
   @ApiResponse({ status: 200, description: 'Return the category.' })
   @ApiResponse({ status: 404, description: 'Category not found.' })
   async findOne(@Param('id') id: string) {
-    return await this.categoryService.findOne(+id);
+    const category = await this.categoryService.findOne(+id);
+
+    if (!category) throw new NotFoundException('Category not found');
+
+    return category;
   }
 
   @Get('name/:name')
@@ -69,6 +77,13 @@ export class CategoryController {
   @ApiResponse({ status: 404, description: 'Category not found.' })
   @HttpCode(204)
   async remove(@Param('id') id: string) {
+
+    const products = await this.productService.findAllByCategory(+id);
+
+    if (products.length > 0) {
+      return { message: 'Category has products. Cannot delete.' };
+    }
+
     const category = await this.categoryService.remove(+id);
 
     if (!category) throw new NotFoundException('Category not found');
